refactor: rename statusbar.ts -> statusBar.ts
