+++ src/statusBar.ts
